+++ pkcs11/tests/basic.rs
@@ -9,9 +9,13 @@ use std::{
 use config_file::{
     CertificateFormat, InstanceConfig, P11Config, RetryConfig, SlotConfig, UserConfig,
 };
-use pkcs11::types::{
-    CKA_MODULUS_BITS, CKA_PUBLIC_EXPONENT, CKA_SIGN, CKA_TOKEN, CKA_VERIFY, CKM_RSA_PKCS,
-    CK_ATTRIBUTE, CK_BBOOL, CK_FALSE, CK_MECHANISM, CK_TRUE, CK_ULONG,
+use pkcs11::{
+    types::{
+        CKA_MODULUS_BITS, CKA_PUBLIC_EXPONENT, CKA_SIGN, CKA_TOKEN, CKA_VERIFY, CKM_RSA_PKCS,
+        CK_ATTRIBUTE, CK_BBOOL, CK_FALSE, CK_MECHANISM, CK_OBJECT_HANDLE, CK_SESSION_HANDLE,
+        CK_TRUE, CK_ULONG,
+    },
+    Ctx,
 };
 
 mod tools;
@@ -106,6 +110,99 @@ fn basic() {
             // Verifying signatures is not supported
             let _signature = ctx.sign(session, &data).unwrap();
             ctx.destroy_object(session, public_key).unwrap();
+            ctx.destroy_object(session, private_key).unwrap();
+        },
+    )
+}
+
+#[test_log::test]
+fn set_attribute_value() {
+    tools::run_tests(
+        &[],
+        P11Config {
+            slots: vec![SlotConfig {
+                label: "Test slot".into(),
+                operator: Some(UserConfig {
+                    username: "operator".into(),
+                    password: Some("opPassphrase".into()),
+                }),
+                administrator: Some(UserConfig {
+                    username: "admin".into(),
+                    password: Some("Administrator".into()),
+                }),
+                description: Some("Test slot".into()),
+                instances: vec![InstanceConfig {
+                    url: "https://localhost:8443/api/v1".into(),
+                    danger_insecure_cert: true,
+                    sha256_fingerprints: Vec::new(),
+                    max_idle_connections: None,
+                }],
+                certificate_format: CertificateFormat::Pem,
+                retries: None,
+                timeout_seconds: Some(10),
+                connections_max_idle_duration: None,
+                tcp_keepalive: None,
+            }],
+            ..Default::default()
+        },
+        |_test_ctx, ctx| {
+            fn get_attributes(
+                ctx: &Ctx,
+                session: CK_SESSION_HANDLE,
+                object: CK_OBJECT_HANDLE,
+            ) -> (String, String) {
+                let mut buffer1 = [0; 128];
+                let mut buffer2 = [0; 128];
+                let mut attributes = vec![
+                    CK_ATTRIBUTE {
+                        attrType: pkcs11::types::CKA_ID,
+                        pValue: buffer1.as_mut_ptr() as _,
+                        ulValueLen: buffer1.len().try_into().unwrap(),
+                    },
+                    CK_ATTRIBUTE {
+                        attrType: pkcs11::types::CKA_LABEL,
+                        pValue: buffer2.as_mut_ptr() as _,
+                        ulValueLen: buffer2.len().try_into().unwrap(),
+                    },
+                ];
+                ctx.get_attribute_value(session, object, &mut attributes)
+                    .unwrap();
+                let id = String::from_utf8(attributes[0].get_bytes().unwrap()).unwrap();
+                let label = String::from_utf8(attributes[1].get_bytes().unwrap()).unwrap();
+                (id, label)
+            }
+
+            let slot = 0;
+            let session = ctx.open_session(slot, 0x04, None, None).unwrap();
+            let (public_key, private_key) = ctx
+                .generate_key_pair(
+                    session,
+                    &RSA_MECHANISM,
+                    RSA_PUBLIC_KEY_ATTRIBUTES,
+                    RSA_PRIVATE_KEY_ATTRIBUTES,
+                )
+                .unwrap();
+
+            let (public_id, public_label) = get_attributes(ctx, session, public_key);
+            println!("public key: id = {public_id}, label = {public_label}");
+            let (private_id, private_label) = get_attributes(ctx, session, private_key);
+            println!("private key: id = {private_id}, label = {private_label}");
+
+            let new_id = "mynewkeyid";
+            ctx.set_attribute_value(
+                session,
+                private_key,
+                &[CK_ATTRIBUTE::new(pkcs11::types::CKA_ID).with_bytes(new_id.as_bytes())],
+            )
+            .unwrap();
+
+            let (public_id, public_label) = get_attributes(ctx, session, public_key);
+            let (private_id, private_label) = get_attributes(ctx, session, private_key);
+            assert_eq!(&public_id, new_id);
+            assert_eq!(&public_label, new_id);
+            assert_eq!(&private_id, new_id);
+            assert_eq!(&private_label, new_id);
+
             ctx.destroy_object(session, private_key).unwrap();
         },
     )