+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Remove support for EC_P224 keys
 - Add support for EC_P256K1, BrainpoolP256, BrainpoolP384 and BrainpoolP512 keys
 - Remove `enable_set_attribute_value` config option
+- Implement `C_SetAttributeValue` for `CKA_ID` to support renaming keys
 
 ## [1.7.2][] (2025-07-25)